requirements.txt
# BDRC Image Database Tool - Python Dependencies

# Database driver
mysql-connector-python>=8.0.0,<9.0.0

# Configuration parsing
PyYAML>=6.0.0,<7.0.0

# Image processing and metadata extraction
Pillow>=10.0.0,<11.0.0
